+++ .env.example
@@ -6,3 +6,5 @@ POSTGRES_VERSION=14
 SERVER_NAME=localhost
 TRUSTED_PROXIES=127.0.0.0/8,10.0.0.0/8,172.16.0.0/12,192.168.0.0/16
 CADDY_MERCURE_JWT_SECRET=JWTSecretKey
+
+BASE_API_URL=https://localhost
