+++ app/src/router/index.ts
@@ -1,7 +1,7 @@
 import { createRouter, createWebHistory } from 'vue-router';
 import TenantHome from '@/views/TenantHome.vue';
 import HomeownerHome from '@/views/HomeowerHome.vue';
-import OwnerView from '../views/OwnerView.vue'
+import HomeownerRegister from '@/views/HomeownerRegister.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -19,7 +19,7 @@ const router = createRouter({
     {
       path: '/homeowner/register',
       name: 'homeowner_register',
-      component: OwnerView,
+      component: HomeownerRegister,
     }
   ]
 })
