--- app/src/views/OwnerView.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-import Owner from '../components/Owner.vue'
-</script>
-
-<template>
-  <main>
-    <Owner />
-  </main>
-</template>
