+++ docker-compose.yml
@@ -89,6 +89,8 @@ services:
       - 4000:4000
     volumes:
       - ./app:/app
+    environment:
+      - VITE_BASE_API_URL=https://localhost
     command: npm run dev
 
   adminer:
